pkg install now optionally waits for marathon.
Since data service applications are fully satisfied by the number of expected
tasks running, there isn't any utility in waiting for the marathon deployment
to complete.  It's theoretically harmless, but cosmos intermittently hiccups
when talking to marathon, so this causes exceptions for the success situation
which are essentially undifferentiable from the failure condition where the
attempt to begin the installation fails.